+++ SimulationCode/src/RandomGenerators.cpp
@@ -2,7 +2,7 @@
 
 RandomGenerators::RandomGenerators() : gen1(rd1()), gen2(rd2()) {}
 
-float RandomGenerators::getRandomFloat(float a, float b) {
+double RandomGenerators::getRandomFloat(double a, double b) {
   std::uniform_real_distribution<> disfloat(a, b);
   return disfloat(gen1);
 }
@@ -23,8 +23,8 @@ template <typename T>
 void RandomGenerators::initializeVelocities(T &velocities, unsigned N) {}
 
 template <>
-void RandomGenerators::initializeVelocities<std::vector<float>>(
-    std::vector<float> &velocities, unsigned numberOfBodies) {
+void RandomGenerators::initializeVelocities<std::vector<double>>(
+    std::vector<double> &velocities, unsigned numberOfBodies) {
   velocities.resize(3 * numberOfBodies);
 
   for (unsigned i = 0; i < numberOfBodies; i++)
@@ -33,6 +33,7 @@ void RandomGenerators::initializeVelocities<std::vector<float>>(
       //velocities[i * 3 + j] = getRandomFloat(-0.01f, 0.01f);
 }
 
+/*
 template <>
 void RandomGenerators::initializeVelocities<std::vector<double>>(
     std::vector<double> &velocities, unsigned numberOfBodies) {
@@ -43,18 +44,19 @@ void RandomGenerators::initializeVelocities<std::vector<double>>(
       velocities[i * 3 + j] = getRandomFloat(-0.01, 0.01);
       //velocities[i * 3 + j] = getRandomFloat(-0.01, 0.01);
 }
+*/
 
 // zderzenie plastyczne / sprezyste
 template <>
-void RandomGenerators::initializeVelocities<thrust::host_vector<float>>(
-    thrust::host_vector<float> &velocities, unsigned numberOfBodies) {
+void RandomGenerators::initializeVelocities<thrust::host_vector<double>>(
+    thrust::host_vector<double> &velocities, unsigned numberOfBodies) {
   velocities.resize(3 * numberOfBodies);
   for (unsigned i = 0; i < numberOfBodies; i++)
     for (int j = 0; j < 3; j++)
       velocities[i * 3 + j] = getRandomFloat(-0.01f, 0.01f);
 }
 
-void RandomGenerators::initializeWeights(std::vector<float> &weights, unsigned numberOfBodies) {
+void RandomGenerators::initializeWeights(std::vector<double> &weights, unsigned numberOfBodies) {
   weights.resize(numberOfBodies);
   int typeMass = 3; //getRandomType();
   unsigned divi;
